requirements.txt
alabaster==0.7.9
altgraph==0.12
Babel==2.3.4
click==6.6
colosseum==0.1.3
docutils==0.12
Flask==0.11.1
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.8
macholib==1.7
MarkupSafe==0.23
modulegraph==0.12.1
py2app==0.10
PyAudio==0.2.9
pydub==0.16.5
Pygments==2.1.3
pyobjc==3.1.1
pyobjc-core==3.1.1
pyobjc-framework-Accounts==3.1.1
pyobjc-framework-AddressBook==3.1.1
pyobjc-framework-AppleScriptKit==3.1.1
pyobjc-framework-AppleScriptObjC==3.1.1
pyobjc-framework-ApplicationServices==3.1.1
pyobjc-framework-Automator==3.1.1
pyobjc-framework-AVFoundation==3.1.1
pyobjc-framework-AVKit==3.1.1
pyobjc-framework-CalendarStore==3.1.1
pyobjc-framework-CFNetwork==3.1.1
pyobjc-framework-CloudKit==3.1.1
pyobjc-framework-Cocoa==3.1.1
pyobjc-framework-Collaboration==3.1.1
pyobjc-framework-Contacts==3.1.1
pyobjc-framework-ContactsUI==3.1.1
pyobjc-framework-CoreBluetooth==3.1.1
pyobjc-framework-CoreData==3.1.1
pyobjc-framework-CoreLocation==3.1.1
pyobjc-framework-CoreText==3.1.1
pyobjc-framework-CoreWLAN==3.1.1
pyobjc-framework-CryptoTokenKit==3.1.1
pyobjc-framework-DictionaryServices==3.1.1
pyobjc-framework-DiskArbitration==3.1.1
pyobjc-framework-EventKit==3.1.1
pyobjc-framework-ExceptionHandling==3.1.1
pyobjc-framework-FinderSync==3.1.1
pyobjc-framework-FSEvents==3.1.1
pyobjc-framework-GameCenter==3.1.1
pyobjc-framework-GameController==3.1.1
pyobjc-framework-ImageCaptureCore==3.1.1
pyobjc-framework-IMServicePlugIn==3.1.1
pyobjc-framework-InputMethodKit==3.1.1
pyobjc-framework-InstallerPlugins==3.1.1
pyobjc-framework-InstantMessage==3.1.1
pyobjc-framework-LatentSemanticMapping==3.1.1
pyobjc-framework-LaunchServices==3.1.1
pyobjc-framework-LocalAuthentication==3.1.1
pyobjc-framework-MapKit==3.1.1
pyobjc-framework-MediaAccessibility==3.1.1
pyobjc-framework-MediaLibrary==3.1.1
pyobjc-framework-MultipeerConnectivity==3.1.1
pyobjc-framework-NetFS==3.1.1
pyobjc-framework-NotificationCenter==3.1.1
pyobjc-framework-OpenDirectory==3.1.1
pyobjc-framework-Photos==3.1.1
pyobjc-framework-PhotosUI==3.1.1
pyobjc-framework-PreferencePanes==3.1.1
pyobjc-framework-PubSub==3.1.1
pyobjc-framework-QTKit==3.1.1
pyobjc-framework-Quartz==3.1.1
pyobjc-framework-SceneKit==3.1.1
pyobjc-framework-ScreenSaver==3.1.1
pyobjc-framework-ScriptingBridge==3.1.1
pyobjc-framework-SearchKit==3.1.1
pyobjc-framework-ServiceManagement==3.1.1
pyobjc-framework-Social==3.1.1
pyobjc-framework-SpriteKit==3.1.1
pyobjc-framework-StoreKit==3.1.1
pyobjc-framework-SyncServices==3.1.1
pyobjc-framework-SystemConfiguration==3.1.1
pyobjc-framework-WebKit==3.1.1
pytz==2016.7
pywebview==1.2.2
rubicon-objc==0.2.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.8
toga==0.2.3
toga-cocoa==0.2.3
toga-core==0.2.3
toga-demo==0.2.3
Werkzeug==0.11.11
